app/requirements.txt
django==3.2.*
djangorestframework==3.12.*
drf-yasg2==1.19.4
django-cors-headers==3.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-decouple==3.3
Pillow==8.2.0
djangorestframework-simplejwt==4.6.0
